+++ index.js
@@ -16,6 +16,7 @@ const make = (name, options) => ([
 
 export const shared = () => [{ dir: join(__dirname, '_shared', 'openapi'), ext: '@' }]
 export const singleUser = options => make('single-user', options)
+export const teamApiTokens = options => make('team-api-tokens', options)
 export const teams = options => make('teams', options)
 export const userApiTokens = options => make('user-api-tokens', options)
 export const userManagement = options => make('user-management', options)
@@ -25,6 +26,7 @@ export const userTasks = options => make('user-tasks', options)
 export const all = options => ([
 	...shared(),
 	...singleUser(options),
+	...teamApiTokens(options),
 	...teams(options),
 	...userApiTokens(options),
 	...userManagement(options),